address feedback: move HandleCompare, test exceptions, reference code source
